refactor: separate prover and verifier to different files
